InsertedCOVID19SpecificRefsToGeneralModelStuff
I just inserted a couple of COVID-19 specific references to the general modeling sentences I previously wrote which was already reviewed. What I wrote was correct for modeling in general so it was useful, but I added here some good specific COVID-19 examples of these general things which could also be useful to readers. (Some of what is already referenced in section may also apply but these references are good to add anyways). Thanks!