SuggestedAddUnderexplordTherapeutics.treatments.md
I put in a paragraph on another type of possible partial-therapeutic/partial-prophylactic that is under clinical trials. I tried to find a place for it that isn't too prominent, but I don't know if I succeeded: For now, at least, I suggest maybe putting it under Underexplored Therapeutics, but maybe it makes more sense as a short subsection earlier in therapeutics? Anyways, I'd appreciate some feedback/comments/review on the paragraph to make sure it's okay to incorporate (& if there's possibly a better place to incorporate it).