RelatedBiologicsWithSeveralTrials.treatments.md
In #349 (2nd file), #408, and on gitterchat we discussed or started PR's and Issues for a few different therapeutics / prophylactics or auxiliary components which are part of several clinical trials: a while ago in gitterchat we mentioned possibly describing the adjuvant components of vaccines (so they are part of many of the vaccine clinical trials), #408 is related to at least on the order of a dozen clinical trials and is also related to types of prophylactics/therapeutics drafted in #349 second file, and convalescent plasma has on the order of 100 trials. Thus, it makes a lot of sense to at least briefly address these. We discussed in #408 and #349 a few different ways to incorporate these. After all that discussion, I thought this way here may be a good one. If people want to fill in a bit more on convalescent plasma, then please do; for now I at least briefly mentioned it (although we mention it briefly in pathogenesis section, it makes sense to also mention in therapeutics). For the other stuff I added a few sentences, but people who wrote vaccine sections may also have opinions on what to mention about adjuvants. For #408 stuff, I incorporated some of the feedback by @rando2, will also look through @rlordan feedback, would like more feedback, and also joined it to related info drafted by @rando2 in #349.