smallAdditionDataExample.diagnostics.md
I mentioned this a bit back in #413 as a useful type of small addition & am adding it now that I had the chance to find a url I was looking for. Here's the explanation/reason why useful:
Towards the end of diagnostics section in the part on Challenges to Diagnostic Approaches / Strategies and Considerations for Determining Whom to Test, there is mention of 1 numerical example of estimated cases versus reported cases. Since this is a variable number, it would be good to add more numerical examples / a range as more data comes out. I would have liked to add this myself, but not quite there yet. Here are two potential examples we could use:
-the 1 referenced example mentioned already in the text right now estimates ~100-fold difference between estimated & reported in late march in Italy (I didn't check if this estimate is based on serology tests or models or both).
-In a pdf file titled NY Forward A GUIDE TO REOPENING NEW YORK & BUILDING BACK BETTER May 2020 which I think I got a while back from the ny.gov website or forward.ny.gov website but can't find it again: It says in May there was ~300,000 cases (~1.5% of NY which has a population of ~20x10^6) reported in NY state and that antibody tests showed ~20% positive. So this is a ~10-fold difference between total-based-on-antibody-tests & reported. But I don't have the full information for this data.
-I found an article on Boise, Idaho [doi:10.1128/JCM.00941-20] which just from browsing the article (haven't assessed it yet) looks like they used ~5000 self-selected individuals for antibody tests and found ~2% in late April. But I don't currently have a corresponding reported cases in ratio to population number to compare this to.
-there's probably more data since I last checked (& will be more later)