ImmuneResponse.pathogenesis.md
Because it was already prepared when we were working on immunology reviewing before, this PR has a simple fast way to cover some of the points specific to SARS-CoV-2 and complements what has already been incorporated in the intro immunology subsection: it starts with a couple of points from standard textbooks (these are unlikely to change anytime soon); then it briefly gives 1 sentence indicators of SARS-CoV-2 reviews/references which cover most of the points in the outline; then I put a basic immunology primer for immunological memory and vaccines because we didn't cover that background yet; finally I point out that more specifics are written in later sections as context for the Diagnostics & Therapeutics.
I don't know if there's another draft for this subsection in the works, but this one can work. The grammar is correct, so we mostly just need a person who has immunology background and read some of the literature to double-check that my brief 1-sentence summaries of the reviews are okay.