finishFixSmallErrorInGoodSuggestion.diagnostics.md
In PR #486 I added a sentence for the reasons described there, but another contributor noticed I had made a typo with "20%" instead of "12%" (thanks!) -- this was partially corrected but along with a great added point which had very slight errors and then we got busy and didn't get to them. So I'm fixing it now as it is an important added point (thanks again!) and by now I have briefly read about even more indications of data at higher resolutions which makes the point even more interesting. What I have now is mathematically & biologically & grammatically correct. Basically, it was suggested also to indicate that different regions/sub-regions had different numbers which was a great suggestion but we didn't finish incorporating it fully correctly -- if you say "up to" or "above" etc. based on this kind of sampling data you are only referring to the specific data and resolution but not what it means about the population at various possible resolutions, so instead I just added a parentheses which explains the general suggested added point with the math/data/modeling lingo. This description is long but the fix is short!