CompletingYEMarshallAuthorUpdate metadata.yaml
This is continuing PR #395 & PR #546 which were written by me about my author information, approved & merged. I am specifying to delete the author information inserted in those files -- i.e., to remove my author information. So that this is clear, I'm putting it as a "new" PR but it is just an extension of the point of the original PR's written by me about my author information. For reasoning, see PR #566 . Thank You!