Nutraceuticals Outline and Checklist
Hi all, here is a proposed checklist of things to update and add to this nutraceutical focused manuscript. I will certainly be able to do this but I would really welcome additions and critical reviews from anyone interested. Particularly the latter as there tends to be a lot of wishful thinking rather than scientific evidence in this field and sometimes when writing about this topic it can be easy to unintentionally re-overstate the claims that these studies make rather than critically review them. Maybe @rando can recommend some reviewers that may be interested as we go.

To add:
1. Abstract to be added
2. A section of misinformation/caution - - I have called this section misinformation for now, but basically just before the conclusions section it is important we highlight and caution the lack of regulation surrounding these products, lack of safety and efficacy data, and misinformation peddled for financial gain. 

To update:
1. Vitamin D - massive strides in research here indicating that patients tend to be vitamin D deficient and there have been two small trials indicating therapeutic benefits when vitamin D is administered early in the disease course.
2. Introduction - update info on sales of nutraceuticals and why this topic is relevant to COVID-19
3. Zinc - some data for zinc in combination with hydroxychloroquine and Azithromycin  has been published
4. Conclusions - tidy up required.

We can add to this checklist as we go, look forward to working with everyone :)