Nutraceuticals Update and re-write
Hi All, 
This is related to #748 and #749, in which part 1-6 are for the most part complete

The nutraceuticals paper has had an update and a substantial facelift.  @rando2 I would definitely welcome some critique and would welcome any additions to the discussion that you may think of.  As aforementioned, it can be difficult to get the tone right with regards to the evidence supporting these products because you don’t want to over hype the claims that are associated with them.
On that topic, the main supplement that actually seems to have significant evidence supporting its potential therapeutic and prophylactic value (and excitement from the nutraceutical community) is vitamin D. I would really appreciate if someone could ensure that the vitamin D section makes sense and is sufficiently critical. 

This is also the first time the paper has had a discussion, conclusion or abstract, so I would welcome additions and feedback. 

There is also a proposed title, it is not very imaginative so if anyone has a better suggestion, go for it :)

I am going to take a break from this paper for a day or two so that I can come back with fresh eyes! I’m sure there will be plenty typos to find then! 

@rando2 mentioned that @hufengling and @marouenbg might also be interested here, but no pressure if you folks are busy. 

Thanks everyone and Happy New Year!