Update to novel variants
Hey @rando2, I added some info on the genetic variants of SARS-CoV-2 and one or two edits. Can you suggest reviewers for this section? I was thinking maybe @marouenbg? I would welcome any and all additions/review of this section. I think it probably needs to be linked better to the previous section for fluidity.  
I also added some details on the clinical consequences of these variants. I hope this is relevant. 

Just on another note, this website has decent graphs of what countries the variants P.1, B.1.315, and B.1.1.7 have been detected. It may be a nice figure. 
https://cov-lineages.org/global_report_B.1.1.7.html

Thanks!