Global vaccine status and distribution
Here is the distribution section - I welcome any and all edits to this section. Some might be irrelevant so feel free to delete if you think it is over explanatory or dealt with elsewhere in the text. This section is also very broad and encompassing and will need an update prior to submission of course as the data will shift rapidly over the next few weeks. I tried to be inclusive as best as the data was available.

We may need to exclude some of the Chinese candidates information if not being included in the main body of the text. Although I might try add those.